Finalized parameter range slider with update and reset buttons
Created (and tested) range slider for parameter value, including capability to manually set the min & max for the map via the slider (but the slider values reset when switching from Licor CO2 to Picarro CO2). Includes two new buttons, one to update the colorbar range based on the values in the slider, and the other one to reset the range to the "global" values.